Renamed and cleaned code
