+++ editor/src/ai/ProviderList/OpenAIProvider.ts
@@ -1,6 +1,6 @@
 // OpenAIProvider.ts
 import axios from 'axios';
-import { IProvider } from './IProvider';
+import { IProvider } from '../ProviderInit/IProvider';
 import { generateRightRegularGrammarPrompt } from '../prompt_templates/rg_prompt';
 import { generateContextFreeGrammarPrompt } from '../prompt_templates/cfg_prompt';
