+++ editor/src/ai/providerConfig.ts
@@ -1,0 +1,3 @@
+import { ProviderType } from "./ProviderInit/ProviderFactory";
+
+export const providerType: ProviderType = ProviderType.OPENAI;
